# factory/factory-service/openapi/factory-openapi.yaml
openapi: 3.1.0
info:
  title: Factory Service API
  description: REST API for the Factory Service — autonomous Kotlin/Spring Boot backend
  version: 0.0.1
servers:
- url: http://localhost:18141
  description: Generated server url
...
paths:
...
  /api/factory/workflows/{workflowId}/environment/release:
    post:
      tags:
      - work-environments
      summary: Release (decommission) the workflow environment
      operationId: releaseWorkUnitEnvironment
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseEnvironmentRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentDataEnvelopeEnvironmentResponse"
  /api/factory/workflows/{workflowId}/environment/reconcile:
    post:
      tags:
      - work-environments
      summary: Reconcile the workflow environment against its Git worktree
      operationId: reconcileWorkUnitEnvironment
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentDataEnvelopeEnvironmentResponse"
  /api/factory/workflows/{workflowId}/environment/provision:
    post:
      tags:
      - work-environments
      summary: Provision (or idempotently reuse) the workflow environment
      operationId: provisionWorkUnitEnvironment
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionEnvironmentRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentDataEnvelopeEnvironmentResponse"
...
  /api/factory/agent-step-questions:
    post:
      tags:
      - agent-step-questions
      summary: Ask a durable human step question and park the attempt in waiting_human.
      description: "Requires an `Authorization: Bearer <capability>` token issued\
        \ by the Factory. Returns as soon as the question is durably recorded — the\
        \ answer never blocks the worker call."
      operationId: askAgentStepQuestion
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      - name: X-AgentOS-Case-Id
        in: header
        required: false
        schema:
          type: string
      - name: X-AgentOS-Agent-Name
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentStepQuestionRequest"
      responses:
        "202":
          description: Question durably recorded (or idempotent replay)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStepQuestionEnvelopeAgentStepQuestionData"
        "400":
          description: "QUESTION_SCHEMA_INVALID, RESULT_IDENTITY_MISMATCH or INVALID_RESULT_REQUEST"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: RESULT_CAPABILITY_INVALID or TRUST_CONTEXT_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: QUESTION_ATTEMPT_NOT_WAITABLE or QUESTION_ALREADY_ASKED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: RESULT_CAPABILITY_EXPIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/factory/workflows/{workflowId}/environment:
    get:
      tags:
      - work-environments
      summary: Get the current environment of a workflow
      operationId: getEnvironmentWorkUnitEnvironment
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentDataEnvelopeEnvironmentResponse"
...
components:
  schemas:
...
    ReleaseEnvironmentRequest:
      type: object
      properties:
        state:
          type: string
    EnvironmentDataEnvelopeEnvironmentResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/EnvironmentResponse"
      required:
      - data
    EnvironmentResponse:
      type: object
      properties:
        revision:
          type: integer
          format: int32
        environment:
          $ref: "#/components/schemas/WorkEnvironment"
        reconciliation:
          $ref: "#/components/schemas/ReconciliationResponse"
        headCommit:
          type: string
        fileAccess:
          $ref: "#/components/schemas/FileAccessResponse"
      required:
      - environment
      - fileAccess
      - revision
    FileAccessResponse:
      type: object
      properties:
        status:
          type: string
        code:
          type: string
        rootPath:
          type: string
      required:
      - rootPath
      - status
    ReconciliationResponse:
      type: object
      properties:
        status:
          type: string
        headCommit:
          type: string
        baseCommit:
          type: string
      required:
      - status
    WorkEnvironment:
      type: object
      properties:
        organizationId:
          type: string
        workstreamId:
          type: string
        environmentId:
          type: string
        workUnitId:
          type: string
        workflowId:
          type: string
        namespaceId:
          type: string
        parentCaseId:
          type: string
        repoRoot:
          type: string
        integrationBranch:
          type: string
        branch:
          type: string
        worktreePath:
          type: string
        baseCommit:
          type: string
        lifecycleState:
          type: string
          enum:
          - provisioning
          - ready
          - busy
          - decommissioned
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        revision:
          type: integer
          format: int32
        envType:
          type: string
      required:
      - branch
      - createdAt
      - createdBy
      - envType
      - environmentId
      - integrationBranch
      - lifecycleState
      - namespaceId
      - organizationId
      - repoRoot
      - revision
      - workUnitId
      - workflowId
      - workstreamId
      - worktreePath
    ProvisionEnvironmentRequest:
      type: object
      properties:
        workUnitId:
          type: string
        integrationBranch:
          type: string
        branch:
          type: string
        namespaceId:
          type: string
        parentCaseId:
          type: string
        repoRoot:
          type: string
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
      required:
      - code
      - message
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      required:
      - error
...
    JsonNode: {}
...
    AgentStepQuestionRequest:
      type: object
      properties:
        attemptId:
          type: string
        question:
          $ref: "#/components/schemas/JsonNode"
    AgentStepQuestionData:
      type: object
      properties:
        attemptId:
          type: string
        interactionId:
          type: string
        status:
          type: string
        idempotent:
          type: boolean
      required:
      - attemptId
      - idempotent
      - interactionId
      - status
    AgentStepQuestionEnvelopeAgentStepQuestionData:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/AgentStepQuestionData"
      required:
      - data
...